tagging 1.4.4
git-svn-id: https://plugins.svn.wordpress.org/comicpress-manager/tags/1.4.4@98426 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ actions/comicpress_build-storyline-schema.php
@@ -13,8 +13,18 @@ function cpm_action_build_storyline_schema() {
 
     $categories_to_create = array();
     $categories_to_rename = array();
+    $category_ids_to_clean = array();
 
     extract(cpm_get_all_comic_categories());
+
+    $comic_posts = cpm_query_posts();
+    $comic_posts_by_category_id = array();
+    foreach ($comic_posts as $post) {
+      foreach (wp_get_post_categories($post->ID) as $category) {
+        if (!isset($comic_posts_by_category_id[$category])) { $comic_posts_by_category_id[$category] = array(); }
+        $comic_posts_by_category_id[$category][] = $post->ID;
+      }
+    }
 
     foreach ($_POST as $field => $value) {
       $parts = explode("/", $field);
@@ -26,6 +36,8 @@ function cpm_action_build_storyline_schema() {
             $cpm_config->messages[] = sprintf(__('Category <strong>%1$s</strong> renamed to <strong>%2$s</strong>.', 'comicpress-manager'), $category['cat_name'], $value);
             $category['cat_name'] = $value;
             wp_update_category($category);
+
+            $category_ids_to_clean[] = $category_id;
           }
         } else {
           $categories_to_create[$field] = $value;
@@ -37,31 +49,39 @@ function cpm_action_build_storyline_schema() {
       }
     }
 
-    foreach ($category_tree as $node) {
-      $category_id = end(explode("/", $node));
+    if (isset($_POST['original-categories'])) {
+      foreach (explode(",", $_POST['original-categories']) as $node) {
+        if (!isset($_POST[$node])) {
+          $category_id = end(explode("/", $node));
+          $category = get_category($category_id);
+          $original_cat_name = $category->cat_name;
 
-      // ensure that we're not deleting a ComicPress category
-      $ok = true;
-      foreach (array('comiccat', 'blogcat') as $type) {
-        if ($category_id == $cpm_config->properties[$type]) { $ok = false; }
-      }
-
-      // ensure that the category truly is a child of the comic category
-      if ($ok) {
-        $category = get_category($category_id);
-        $ok = false;
-
-        if (!is_wp_error($category)) {
-          while (($category->parent != 0) && ($category->parent != $cpm_config->properties['comiccat'])) {
-            $category = get_category($category->parent);
+          // ensure that we're not deleting a ComicPress category
+          $ok = true;
+          foreach (array('comiccat', 'blogcat') as $type) {
+            if ($category_id == $cpm_config->properties[$type]) { $ok = false; }
           }
-          if ($category->parent == $cpm_config->properties['comiccat']) { $ok = true; }
-        }
-      }
 
-      if ($ok) {
-        wp_delete_category($category_id);
-        $cpm_config->messages[] = sprintf(__('Category <strong>%s</strong> deleted.', 'comicpress-manager'), $category->cat_name);
+          // ensure that the category truly is a child of the comic category
+          if ($ok) {
+            $category = get_category($category_id);
+            $ok = false;
+
+            if (!is_wp_error($category)) {
+              while (($category->parent != 0) && ($category->parent != $cpm_config->properties['comiccat'])) {
+                $category = get_category($category->parent);
+              }
+              if ($category->parent == $cpm_config->properties['comiccat']) { $ok = true; }
+            }
+          }
+
+          if ($ok) {
+            wp_delete_category($category_id);
+            $category_ids_to_clean[] = $category_id;
+
+            $cpm_config->messages[] = sprintf(__('Category <strong>%s</strong> deleted.', 'comicpress-manager'), $original_cat_name);
+          }
+        }
       }
     }
 
@@ -87,6 +107,8 @@ function cpm_action_build_storyline_schema() {
 
       if (!category_exists($value)) {
         $category_id = wp_create_category($value, $parent_id);
+        $category_ids_to_clean[] = $category_id;
+
         array_push($parts, $parent_id);
         array_push($parts, $category_id);
         $changed_field_ids[$original_field] = implode("/", $parts);
@@ -107,16 +129,57 @@ function cpm_action_build_storyline_schema() {
 
     // ensure we're writing sane data
     $new_order = array();
+    $valid_comic_categories = array();
     foreach ($order as $node) {
       $parts = explode("/", $node);
       if (($parts[0] == "0") && (count($parts) > 1)) {
         $new_order[] = $node;
+        $valid_comic_categories[] = end($parts);
       }
+    }
+
+    $comic_categories_preserved = array();
+    foreach ($comic_posts as $post) {
+      $categories = wp_get_post_categories($post->ID);
+      if (count(array_intersect($valid_comic_categories, $categories)) == 0) {
+        $all_parent_categories = array();
+
+        foreach ($comic_posts_by_category_id as $category => $post_ids) {
+          if (in_array($post->ID, $post_ids)) {
+            foreach ($new_order as $node) {
+              $parts = explode("/", $node);
+              if ($category == end($parts)) {
+                $parts = explode("/", $node);
+                array_pop($parts);
+                if (count($parts) > 1) { $all_parent_categories[] = implode("/", $parts); }
+              }
+            }
+          }
+        }
+
+        if (count($all_parent_categories) > 0) {
+          foreach ($all_parent_categories as $category_node) {
+            if (in_array($category_node, $new_order)) {
+              $categories[] = end(explode("/", $category_node));
+            }
+          }
+        } else {
+          $categories[] = $cpm_config->properties['comiccat'];
+        }
+
+        wp_set_post_categories($post->ID, $categories);
+        $comic_categories_preserved[] = $post->ID;
+      }
+    }
+
+    if (count($comic_categories_preserved) > 0) {
+      $cpm_config->messages[] = sprintf(__("The following orphaned comic posts were placed into their original category's parent: <strong>%s</strong>"), implode(", ", $comic_categories_preserved));
     }
 
     $cpm_config->messages[] = __('Storyline structure saved.', 'comicpress-manager');
     update_option("comicpress-storyline-category-order", implode(",", $new_order));
 
+    clean_term_cache($category_ids_to_clean, 'category');
     wp_cache_flush();
   }
 }
